requirements.txt
aerich==0.9.2 ; python_version >= "3.12" and python_version < "3.15"
aiofiles==25.1.0 ; python_version >= "3.12" and python_version < "3.15"
aiogram==3.27.0 ; python_version >= "3.12" and python_version < "3.15"
aiohappyeyeballs==2.6.1 ; python_version >= "3.12" and python_version < "3.15"
aiohttp==3.13.5 ; python_version >= "3.12" and python_version < "3.15"
aiosignal==1.4.0 ; python_version >= "3.12" and python_version < "3.15"
aiosqlite==0.22.1 ; python_version >= "3.12" and python_version < "3.15"
amqp==5.3.1 ; python_version >= "3.12" and python_version < "3.15"
annotated-doc==0.0.4 ; python_version >= "3.12" and python_version < "3.15"
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "3.15"
anyio==4.13.0 ; python_version >= "3.12" and python_version < "3.15"
apscheduler==3.11.2 ; python_version >= "3.12" and python_version < "3.15"
asgiref==3.11.1 ; python_version >= "3.12" and python_version < "3.15"
asyncclick==8.3.0.7 ; python_version >= "3.12" and python_version < "3.15"
asyncpg==0.31.0 ; python_version >= "3.12" and python_version < "3.15"
attrs==26.1.0 ; python_version >= "3.12" and python_version < "3.15"
bcrypt==4.3.0 ; python_version >= "3.12" and python_version < "3.15"
billiard==4.2.4 ; python_version >= "3.12" and python_version < "3.15"
celery-redbeat==2.3.3 ; python_version >= "3.12" and python_version < "3.15"
celery==5.6.3 ; python_version >= "3.12" and python_version < "3.15"
certifi==2026.4.22 ; python_version >= "3.12" and python_version < "3.15"
charset-normalizer==3.4.7 ; python_version >= "3.12" and python_version < "3.15"
click-didyoumean==0.3.1 ; python_version >= "3.12" and python_version < "3.15"
click-plugins==1.1.1.2 ; python_version >= "3.12" and python_version < "3.15"
click-repl==0.3.0 ; python_version >= "3.12" and python_version < "3.15"
click==8.3.3 ; python_version >= "3.12" and python_version < "3.15"
cffi==2.0.0 ; python_version >= "3.12" and python_version < "3.15"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "3.15" and (platform_system == "Windows" or sys_platform == "win32")
cryptography==46.0.5 ; python_version >= "3.12" and python_version < "3.15"
deprecated==1.3.1 ; python_version >= "3.12" and python_version < "3.15"
dictdiffer==0.9.0 ; python_version >= "3.12" and python_version < "3.15"
distro==1.9.0 ; python_version >= "3.12" and python_version < "3.15"
dnspython==2.8.0 ; python_version >= "3.12" and python_version < "3.15"
email-validator==2.3.0 ; python_version >= "3.12" and python_version < "3.15"
fastadmin==0.2.22 ; python_version >= "3.12" and python_version < "3.15"
fastapi==0.136.1 ; python_version >= "3.12" and python_version < "3.15"
frozenlist==1.8.0 ; python_version >= "3.12" and python_version < "3.15"
h11==0.16.0 ; python_version >= "3.12" and python_version < "3.15"
httpcore==1.0.9 ; python_version >= "3.12" and python_version < "3.15"
httpx==0.28.1 ; python_version >= "3.12" and python_version < "3.15"
idna==3.13 ; python_version >= "3.12" and python_version < "3.15"
iso8601==2.1.0 ; python_version >= "3.12" and python_version < "3.15"
kombu==5.6.2 ; python_version >= "3.12" and python_version < "3.15"
langdetect==1.0.9 ; python_version >= "3.12" and python_version < "3.15"
loguru==0.7.3 ; python_version >= "3.12" and python_version < "3.15"
magic-filter==1.0.12 ; python_version >= "3.12" and python_version < "3.15"
multidict==6.7.1 ; python_version >= "3.12" and python_version < "3.15"
netaddr==1.3.0 ; python_version >= "3.12" and python_version < "3.15"
packaging==26.1 ; python_version >= "3.12" and python_version < "3.15"
prompt-toolkit==3.0.52 ; python_version >= "3.12" and python_version < "3.15"
propcache==0.4.1 ; python_version >= "3.12" and python_version < "3.15"
pydantic-core==2.41.5 ; python_version >= "3.12" and python_version < "3.15"
pydantic-settings==2.14.0 ; python_version >= "3.12" and python_version < "3.15"
pydantic==2.12.5 ; python_version >= "3.12" and python_version < "3.15"
pycparser==2.23 ; python_version >= "3.12" and python_version < "3.15"
pyjwt==2.12.1 ; python_version >= "3.12" and python_version < "3.15"
pypika-tortoise==0.6.5 ; python_version >= "3.12" and python_version < "3.15"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "3.15"
python-dotenv==1.2.2 ; python_version >= "3.12" and python_version < "3.15"
pytz==2026.1.post1 ; python_version >= "3.12" and python_version < "3.15"
redis==4.6.0 ; python_version >= "3.12" and python_version < "3.15"
requests==2.33.1 ; python_version >= "3.12" and python_version < "3.15"
six==1.17.0 ; python_version >= "3.12" and python_version < "3.15"
starlette==1.0.0 ; python_version >= "3.12" and python_version < "3.15"
tenacity==9.1.4 ; python_version >= "3.12" and python_version < "3.15"
tortoise-orm==0.25.4 ; python_version >= "3.12" and python_version < "3.15"
typing-extensions==4.15.0 ; python_version >= "3.12" and python_version < "3.15"
typing-inspection==0.4.2 ; python_version >= "3.12" and python_version < "3.15"
tzdata==2026.1 ; python_version >= "3.12" and python_version < "3.15"
tzlocal==5.3.1 ; python_version >= "3.12" and python_version < "3.15"
urllib3==2.6.3 ; python_version >= "3.12" and python_version < "3.15"
uvicorn==0.46.0 ; python_version >= "3.12" and python_version < "3.15"
vine==5.1.0 ; python_version >= "3.12" and python_version < "3.15"
wcwidth==0.6.0 ; python_version >= "3.12" and python_version < "3.15"
websockets==16.0 ; python_version >= "3.12" and python_version < "3.15"
win32-setctime==1.2.0 ; python_version >= "3.12" and python_version < "3.15" and sys_platform == "win32"
wrapt==2.1.2 ; python_version >= "3.12" and python_version < "3.15"
yarl==1.23.0 ; python_version >= "3.12" and python_version < "3.15"
yookassa==2.5.0 ; python_version >= "3.12" and python_version < "3.15"
